rnr form line to own file
